requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
Babel==2.7.0
black==19.3b0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.39.1
django-database-view==0.1.3
django-debug-toolbar==2.0
django-extensions==2.1.9
django-hosts==3.0
django-location-field==2.1.0
django-sslserver==0.20
djangorestframework==3.9.4
geographiclib==1.49
geopy==1.20.0
google-auth==1.6.3
gunicorn==19.9.0
idna==2.8
isort==4.3.18
lazy-object-proxy==1.4.0
mccabe==0.6.1
oauthlib==3.0.1
packaging==19.0
Pillow==6.0.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.3.1
pyparsing==2.4.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.3.5
urllib3==1.25.3
virtualenv-clone==0.5.3
whitenoise==4.1.3
wincertstore==0.2
wrapt==1.11.1
